Use send instead of eval for :load'ing
The difference between `eval` and `send` is that `eval` captures the
output of the command, prints it, then asks you to hit ENTER. `send`
just carries right on.

Since the output from `:load`ing a module is usually just chunder, (and
since it gets duplicated into the REPL anyway), we should just use
`send`.