Prevent infinite loop on :InteroOpen
When :InteroOpen is run before :InteroStart, it can get into an infinite
loop:

- calls `start()`
- `start()` says "Intero is still compiling"
- `start()` exits and returns to `open()`
- `open()` thinks we're started now, and recurses so we can find the
  buffer information
- nothing changes, and this repeats ad infinitum

This bug arose because of the option to delay starting Intero
immediately for every Haskell file. We used to be assuming a synchronous
initialization, preventing anything else from happening in the meantime.
Now, the initialization happens asynchronously, and we have to deal with
that.