Improve documentation (#53)
* Add online help (:help intero)

Intero deserves a fancy help file! It makes the project look more
attractive to new users. Also contains answers to a bunch of questions I
had when first starting out.

* Update README

With the new help doc, we probably don't wan to duplicate documentation
too much. But, we still want the README to be useful.

This new README has the essentials for a casual GitHub browser, and
defers to the documentation for everything else.

Also, has a fancy new screencast!

* fixup: Add CI badge

I had removed this at first because it wasn't being used (Travis wasn't
even configured for this repo). It looks like that's changed, so let's
add it back!