Parse GHCi version on repl startup (#115)
This is in preparation for supporting any GHCi, not only intero. We'll
need version checks to switch some features on and off, e.g. `:type` and
`:type-at` usage.

In itself, this commit shouldn't affect the behaviour, but it does some
refactoring, so please review changes to the `process.vim` file.

This commit does not add support for using something other than
`intero`.