Also trigger Neomake errors from :load (#151)
Before, we would only update the Neomake error list when using
`:reload`. My workflow uses `:load` a lot; it'd be nice to also get the
Neomake errors populated when using one of these commands.

I've been using this change locally for a few weeks without error, so it
should be pretty safe.